backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["dhythm <yuta.okada.20@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pygments = "^2.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pylint = "^2.16.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
